# Remove recursion in KBucket