This PR demonstrate idea from #34 
if you ok with this, I'll fix tests