@tristanls 2016-05-20
Ok, I see what the nature of the change is now, thank you for showing it via code.

I can see a way to benchmark the speed. I'm still uncertain how to profile lower memory. 

My assumption, based on comments in #34, is that this pull request will improve speed, memory consumption, or both, without making any worse. It's fine to me if the speed and/or memory criteria are met.