@fanatid 2016-08-09
Updated. Tests are fixed now.
@tristanls is benchmark required for merge this PR? If so, I'll try write something (but I think, numbers can be almost the same influenced by jitter).