@tristanls 2016-08-14
@fanatid, welcome back :)

Without some sort of way to measure the difference, why is this an improvement? If it's not an improvement, why do anything?