@fanatid 2016-08-14
Benchmark for `KBuket.add`:
```bash
(old)
$ SEED=944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001 node add.js 
Seed: 944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001
KBucket.add: 930.976ms
Memory:  { rss: 37056512, heapTotal: 15724544, heapUsed: 5543600 }
$ SEED=944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001 node add.js 
Seed: 944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001
KBucket.add: 918.060ms
Memory:  { rss: 36798464, heapTotal: 15724544, heapUsed: 7983504 }
$ SEED=944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001 node add.js 
Seed: 944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001
KBucket.add: 929.083ms
Memory:  { rss: 36950016, heapTotal: 15724544, heapUsed: 8001008 }

(new)
$ SEED=944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001 node add.js 
Seed: 944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001
KBucket.add: 919.014ms
Memory:  { rss: 37085184, heapTotal: 15724544, heapUsed: 6294584 }
$ SEED=944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001 node add.js 
Seed: 944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001
KBucket.add: 921.948ms
Memory:  { rss: 36995072, heapTotal: 15724544, heapUsed: 6302752 }
$ SEED=944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001 node add.js 
Seed: 944f41e2735008807c0a023244042dc6bba49b971cb983d252c61f348a959001
KBucket.add: 862.595ms
Memory:  { rss: 37371904, heapTotal: 15724544, heapUsed: 6308344 }
```