@tristanls 2016-08-14
Agreed, I see similar behavior on my machine.

```
$ node -v
v6.3.1

[k-bucket (feature/non-recursive %=)]$ node _tmp/_add.js 
Seed: a9b220adad92e1ecdd014c7582959942363909e1929b4a580292c01538dd072b
KBucket.add: 903.401ms
Memory:  { rss: 34217984, heapTotal: 15724544, heapUsed: 6305320 }
[k-bucket (feature/non-recursive %=)]$ node _tmp/_add.js 
Seed: 336f4b6e8ca073ee803ff753ad00a024cf1218c905c92b7c2d0e6f4a979464ed
KBucket.add: 871.725ms
Memory:  { rss: 34127872, heapTotal: 15724544, heapUsed: 6301776 }
[k-bucket (feature/non-recursive %=)]$ node _tmp/_add.js 
Seed: 5cfea93f7323fa32169a9b40f3c0cc839220514c2d5fa7e10c46f36182ecaf5d
KBucket.add: 856.090ms
Memory:  { rss: 35250176, heapTotal: 15724544, heapUsed: 5851992 }
[k-bucket (feature/non-recursive %=)]$ node _tmp/_add.js 
Seed: 3fdcf1e2971facad2423b070026c8038858447d714f09cad4d90d5aa433c85d5
KBucket.add: 860.850ms
Memory:  { rss: 34304000, heapTotal: 15724544, heapUsed: 5868760 }
[k-bucket (feature/non-recursive %=)]$ node _tmp/_add.js 
Seed: c40f5faddd703631d3be6a8e97e69e0ace46cf8261f47d3d5a02efefd03bb874
KBucket.add: 860.164ms
Memory:  { rss: 35237888, heapTotal: 15724544, heapUsed: 5870328 }
[k-bucket (feature/non-recursive %=)]$ node _tmp/_add.js 
Seed: eab42f99fbfbaaf01040a25fb0c4d63b4066775e6e49c23c5858da3c2a77f34a
KBucket.add: 860.007ms
Memory:  { rss: 35631104, heapTotal: 15724544, heapUsed: 6296016 }
[k-bucket (feature/non-recursive %=)]$ git checkout master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
[k-bucket (master %=)]$ node _tmp/_add.js 
Seed: d214701fd5f1481ff9b6d9668cc2c32aaeebd01b0b0f4c47e8bf1ab081ea944e
KBucket.add: 871.814ms
Memory:  { rss: 35389440, heapTotal: 15724544, heapUsed: 7999312 }
[k-bucket (master %=)]$ node _tmp/_add.js 
Seed: 42f49bace0a9d41960f66feb98b4f1105fe93dc15f055a21cb2ed7651056fd09
KBucket.add: 862.559ms
Memory:  { rss: 36229120, heapTotal: 15724544, heapUsed: 8026056 }
[k-bucket (master %=)]$ node _tmp/_add.js 
Seed: 41de7d30beb4cbf36aacba01aa885929951e5ba1e1af76ab7e12a2e7a35ee731
KBucket.add: 879.634ms
Memory:  { rss: 35143680, heapTotal: 15724544, heapUsed: 7994664 }
[k-bucket (master %=)]$ node _tmp/_add.js 
Seed: c1b556f3bae66745f18d53ba6e4f3f8bc0a8f782e4b5e707fc021613476c24d0
KBucket.add: 877.387ms
Memory:  { rss: 34439168, heapTotal: 15724544, heapUsed: 7981200 }
[k-bucket (master %=)]$ node _tmp/_add.js 
Seed: a48674476ebfcbbfb06b52ac5c9701cf59d0b10bb36e6d7bb0cccd8fa82833d4
KBucket.add: 884.902ms
Memory:  { rss: 35414016, heapTotal: 15724544, heapUsed: 8007912 }
```