@fanatid 2016-08-14
Almost same time, but memory consumption is better. Are you going to merge this PR? :)