@feross 2016-08-21
It appears that this PR changed behavior in a way that affected bittorrent-dht: https://github.com/tristanls/k-bucket/issues/38